+++ Userland/Libraries/LibGUI/Layout.h
@@ -9,25 +9,25 @@
 #include <AK/OwnPtr.h>
 #include <AK/Vector.h>
 #include <AK/WeakPtr.h>
-#include <LibCore/Object.h>
 #include <LibGUI/Forward.h>
 #include <LibGUI/Margins.h>
+#include <LibGUI/Object.h>
 #include <LibGUI/UIDimensions.h>
 #include <LibGfx/Forward.h>
 
-namespace Core::Registration {
-extern Core::ObjectClassRegistration registration_Layout;
+namespace GUI::Registration {
+extern GUI::ObjectClassRegistration registration_Layout;
 }
 
-#define REGISTER_LAYOUT(namespace_, class_name)                                                                                                       \
-    namespace Core::Registration {                                                                                                                    \
-    Core::ObjectClassRegistration registration_##class_name(                                                                                          \
-        #namespace_ "::" #class_name##sv, []() { return static_ptr_cast<Core::Object>(namespace_::class_name::construct()); }, &registration_Layout); \
+#define REGISTER_LAYOUT(namespace_, class_name)                                                                                                      \
+    namespace GUI::Registration {                                                                                                                    \
+    ::GUI::ObjectClassRegistration registration_##class_name(                                                                                        \
+        #namespace_ "::" #class_name##sv, []() { return static_ptr_cast<GUI::Object>(namespace_::class_name::construct()); }, &registration_Layout); \
     }
 
 namespace GUI {
 
-class Layout : public Core::Object {
+class Layout : public GUI::Object {
     C_OBJECT_ABSTRACT(Layout);
 
 public:
