+++ Userland/Libraries/LibGUI/ScrollableContainerWidget.cpp
@@ -130,7 +130,7 @@ ErrorOr<void> ScrollableContainerWidget::load_from_gml_ast(NonnullRefPtr<GUI::GM
         auto class_name = content_widget.name();
 
         RefPtr<Core::Object> child;
-        if (auto* registration = Core::ObjectClassRegistration::find(class_name)) {
+        if (auto* registration = GUI::ObjectClassRegistration::find(class_name)) {
             child = TRY(registration->construct());
         } else {
             child = TRY(unregistered_child_handler(class_name));
